requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.5
django-rest-framework==0.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
MouseInfo==0.1.3
Pillow==7.2.0
PyAutoGUI==0.9.50
PyGetWindow==0.0.8
PyJWT==2.4.0
PyMsgBox==1.0.8
pyperclip==1.8.0
PyRect==0.1.4
PyScreeze==0.1.26
PyTweening==1.0.3
pytz==2022.1
selenium==3.141.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.25.9
